refactor: Restructure project and rename packages
