+++ src/main/java/com/wizeflow/crm/infrastructure/entity/Appointment.java
