+++ src/main/java/com/wizeflow/crm/infrastructure/entity/ChatSession.java
