+++ src/main/java/com/wizeflow/crm/infrastructure/entity/Lead.java
