+++ src/main/java/com/wizeflow/crm/infrastructure/entity/N8nChatHistories.java
