+++ src/main/java/com/wizeflow/crm/infrastructure/entity/Ticket.java
