--- src/main/resources/application-dev.properties
@@ -1,15 +1,0 @@
-# -- Profile: dev (Supabase PostgreSQL) --
-# Requires .env file with DB_URL, DB_USERNAME, DB_PASSWORD
-
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-
-# -- HikariCP - SSL required for Supabase --
-spring.datasource.hikari.data-source-properties.sslmode=require
-spring.datasource.hikari.maximum-pool-size=5
-
-# -- Logging --
-logging.level.org.hibernate.SQL=DEBUG
-logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
-logging.level.com.wizeflow=DEBUG
