--- src/main/resources/application-prod.properties
@@ -1,16 +1,0 @@
-# -- Profile: prod (Self-hosted PostgreSQL) --
-# Activate with: SPRING_PROFILES_ACTIVE=prod
-# Requires env vars: DB_URL, DB_USERNAME, DB_PASSWORD
-
-spring.jpa.hibernate.ddl-auto=validate
-spring.jpa.show-sql=false
-spring.jpa.properties.hibernate.format_sql=false
-
-# -- HikariCP --
-spring.datasource.hikari.data-source-properties.sslmode=require
-spring.datasource.hikari.maximum-pool-size=10
-
-# -- Logging --
-logging.level.org.hibernate.SQL=WARN
-logging.level.com.wizeflow=INFO
-
