+++ src/test/java/com/wizeflow/crm/WizeFlowCrmApplicationTests.java
@@ -1,12 +1,10 @@
-package com.wizeflow.crm_backend;
+package com.wizeflow.crm;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
-@ActiveProfiles("test")
-class CrmBackendApplicationTests {
+class WizeFlowCrmApplicationTests {
 
 	@Test
 	void contextLoads() {
